Db changes and new screen

+++ CncDataSaver.BL/Model/Channel2.cs
@@ -10,20 +10,20 @@ namespace CncDataSaver.BL.Model
     {
         public int Channel2Id { get; set; }
         public virtual ICollection<OffsetData> Offsets { get; set; }
-        public double? X1 { get; set; }
-        public double? Z1 { get; set; }
-        public double? X2 { get; set; }
-        public double? Z2 { get; set; }
-        public double? X3 { get; set; }
-        public double? Z3 { get; set; }
-        public double? X4 { get; set; }
-        public double? Z4 { get; set; }
-        public double? X5 { get; set; }
-        public double? Z5 { get; set; }
-        public double? X6 { get; set; }
-        public double? Z6 { get; set; }
-        public double? X7 { get; set; }
-        public double? Z7 { get; set; }
+        public decimal? X1 { get; set; }
+        public decimal? Z1 { get; set; }
+        public decimal? X2 { get; set; }
+        public decimal? Z2 { get; set; }
+        public decimal? X3 { get; set; }
+        public decimal? Z3 { get; set; }
+        public decimal? X4 { get; set; }
+        public decimal? Z4 { get; set; }
+        public decimal? X5 { get; set; }
+        public decimal? Z5 { get; set; }
+        public decimal? X6 { get; set; }
+        public decimal? Z6 { get; set; }
+        public decimal? X7 { get; set; }
+        public decimal? Z7 { get; set; }
 
         public override string ToString()
         {
